+ Added MaterialPictureBox +Added a new method of Opacity, the following three objects now have it and will be forward to the rest of them later + Added Opacity+ to MaterialPictureBox + Added Opacity+ to MaterialLabel + Added Opacity+ to Button
Opacity+ is actually very simple and should be the way we added opacity in our controls before.

Before we used to change the opacity of the font color, back color, border color, etc. This is why we didn't port it to more complex controls like the TabHeader. Now, we just draw the back color of the parent (or what ever back color you want) over the control. Make sure to run Control.UpdateOpacityColor() for each control since this can't be start up currently for controls since the parent isn't initialized yet.